Added service interface
Added service interface and created impl class to make it re usable and extend-able
