+++ src/main/java/com/example/unitconverter/api/controllers/ConversionController.java
@@ -1,7 +1,5 @@
 package com.example.unitconverter.api.controllers;
 
-import com.example.unitconverter.api.dtos.ConversionResponseDto;
-import com.example.unitconverter.service.UnitConversionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,6 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.example.unitconverter.api.dtos.ConversionResponseDto;
+import com.example.unitconverter.service.UnitConversionService;
 
 @RestController
 @RequestMapping("/convert")
